SPU-3962

+++ lib/features/convert_flow/screens/convert_tab_body.dart
@@ -94,7 +94,7 @@ class _BuyAmountScreenBodyState extends State<ConvertAmountTabBody> with Automat
                     onPaste: () async {
                       final data = await Clipboard.getData('text/plain');
                       if (data?.text != null) {
-                        final n = int.tryParse(data!.text!);
+                        final n = double.tryParse(data!.text!);
                         if (n != null) {
                           store.pasteValue(n.toString().trim());
                         }
