+++ lib/features/currency_sell/ui/currency_sell.dart
@@ -175,7 +175,7 @@ class _CurrencySellBodyState extends State<CurrencySellBody> {
               onPaste: () async {
                 final data = await Clipboard.getData('text/plain');
                 if (data?.text != null) {
-                  final n = int.tryParse(data!.text!);
+                  final n = double.tryParse(data!.text!);
                   if (n != null) {
                     store.pasteValue(n.toString().trim());
                   }
