+++ lib/features/iban/iban_send/iban_send_amount/ui/iban_send_amount.dart
@@ -117,7 +117,7 @@ class IbanSendAmountBody extends StatelessObserverWidget {
                   onPaste: () async {
                     final data = await Clipboard.getData('text/plain');
                     if (data?.text != null) {
-                      final n = int.tryParse(data!.text!);
+                      final n = double.tryParse(data!.text!);
                       if (n != null) {
                         store.pasteAmount(n.toString().trim());
                       }
