+++ lib/features/withdrawal/send_card_detail/send_globally_amount_screen.dart
@@ -125,7 +125,7 @@ class _SendGloballyAmountScreenBodyState extends State<SendGloballyAmountScreenB
               onPaste: () async {
                 final data = await Clipboard.getData('text/plain');
                 if (data?.text != null) {
-                  final n = int.tryParse(data!.text!);
+                  final n = double.tryParse(data!.text!);
                   if (n != null) {
                     store.pasteAmount(n.toString().trim());
                   }
